pyproject.toml
[tool.poetry]
name = "telegram-auth"
version = "0.1.0"
description = ""
authors = ["serebriakovanton <cu-hawk@yandex.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.3"
psycopg2-binary = "^2.9.10"
python-dotenv = "^1.0.1"
python-telegram-bot = "^21.8"



[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
